fix: correct off by 1 on cache size calculation (#86)
* fix: correct off by 1 on cache size calculation
fix: update inconsistent naming of 'rows_to_discard'
feat: add some cache calculation sanity checks (test)

* feat: let 'rows_to_discard' be optional in gen_cached_proof

The idea is that if it's specified, it will be used, even in the case
that configuration is bad (allowing the caller to tune it).  If not
specified, a reasonable default will be chosen.  Tests should
generally allow the default to be used, and production code can
specify as needed.

* fix: rename incorrectly named variable

* fix: add comment & rename incorrectly named 'height' usage